+++ tests/promises/fixes.validate.js
@@ -6,7 +6,7 @@ export default ($) => {
     it('$L validate() should be false', (done) => {
       $.$L.validate().then(({ isValid }) => {
         describe('Form $L checks after validate()', () => {
-          it('$L state.option "validateOnChange" should be true', () =>
+          it('$L state.options "validateOnChange" should be true', () =>
             expect($.$L.state.options.get('validateOnChange')).to.be.true);
 
           it('$L email hasError should be true', () =>